Fix that should prevent players from learning more than two primary professions.
I cannot be certain how the hack that let a player learn all professions was performed, but my best guess is that is was done in the same way as the guild creation hack.

I have added an extra check in the function that makes the trainer cast the teaching spell that should ensure a teacher never teaches more than the maximum amount of primary professions.