More work on threaded movement handling.
Reworked the movement threading to use the thread pool instead of one
thread per player. This should give better performance as context
switching will hopefully be less.

Also added in an extra mutex that protects the grid from corruption if
the movement worker threads need to relocate a player to another grid.