Stability fix in the movement opcode worker.
There was a rare problem that could arise if a player unloaded or
changed map while a mutex in the update function was held. This would
cause the map pointer to get set to 0 and resulting in a crash later.

Remidied this by saving the map pointer at the beginning of the function
and using the same pointer throughout the function's runtime.