Removed obsolete buggy threading code.
Removed the threading code that was supposed to keep players from
lagging around.

The issue wasn't that the server wasn't powerful enough
to handle the updates but how timestamps for movement were handled.
With correct timestamps server performance shouldn't play a large part
and the code is mostly causing problems.