Players' positions will update while in the air.
Players will now have their position on the server updated while
they're ni the air.

There was code written to actively prevent this from happening. Since
we can't think of any immediate ill effects and players are asking for
it the code that prevents the updates has been removed.

If things start behaving strangely this might be something to look into.