Characters highest rank should no appear correctly.
There was an issue with characters with over 25 life-time kills that
have never been in the standings not getting the correct highest rank
set.

All characters with more than 25, or whatever is set in the config
file, kills should now get their highest rank set to rank 1 (5 in the
database).