Mobs can now ignore the non-attackable flag.
The mobs in the RP for "A Pawn on the Eternal Board" should attack
eachother, but the player should be unable to attack them. To make this
possible I added the method Creature::IgnoreNonAttackableFlag which,
when set to true, will make mobs able to attack other mobs with the
non-attackable flag.