Mind Control - movement packets should be correctly ignored.
There has long been an issue with mind controlling creatures seemingly
bugging around. It has been most notable for the MC totems in ZG and
Hakkar.

The issue was that when the players control over its character is
disabled movement packages might still be on their way to the server.
Since the mover is set to the creature that is controlling the player
this caused the mindcontrolled creature to start moving around.